pyproject.toml
[tool.poetry]
name = "h6"
version = "0.1.1"
description = "A Python library to emulate a Zoom H6 recorder remote control"
authors = ["Matias Godoy <mattogodoy@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/mattogodoy/h6/"
repository = "https://github.com/mattogodoy/h6/"
keywords = ["zoom", "h6", "remote"]

[tool.poetry.dependencies]
python = "^3.6"
click = "^7.1.2"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"

[tool.poetry.scripts]
h6 = 'h6.cli:send_command'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
